requirements.txt
# Flask Backend Dependencies
Flask>=2.3.0
Flask-CORS>=4.0.0
requests>=2.25.0
python-dotenv>=1.0.0

# Data Science and Analysis
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0

# Space/Astronomy Libraries
astropy>=5.0
skyfield>=1.40
ephem>=4.0

# API and Data Processing
httpx>=0.24.0

# Development Tools
jupyter>=1.0.0
pytest>=6.0.0
black>=22.0.0
flake8>=4.0.0
